requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.6
django-allauth==0.41.0
django-import-export==2.5.0
djangorestframework==3.10.0
et-xmlfile==1.1.0
idna==2.10
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
